requirements_mqtt.txt
Flask==2.3.3
Flask-CORS==4.0.0
paho-mqtt==1.6.1
gunicorn==21.2.0 